Move settings to new index file
